requirements.txt
# Intelligent Safety Net — Phase 0 Dependencies

# OpenAI API
openai>=1.0

# Data handling
pandas>=2.0
openpyxl>=3.1

# PDF processing
pdf2image>=1.16
Pillow>=10.0

# Environment
python-dotenv>=1.0

# UI
streamlit>=1.30

# Fuzzy matching (future use)
fuzzywuzzy[speedup]>=0.18
python-Levenshtein>=0.21
